+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - All horizontal column alignments will be rendered according to the table config
 
+## Additions
+
+- Add disable prop to StandardTableHtmlCopyToClipboardButton
+
 ## 9.0.0
 
 ### Breaking changes